- Add ignored words in spelling - Add period to help strings - Fix locking problem in FlieHook (lock not released on error) - Change Formatter to handle Color properly - Fix GenericHook.Levels to be more efficient and flexible (accepting level higher than existing levels) - Reorganize files (level.go, multi_logger.go)
