+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+{
+    "cSpell.words": [
+        "logrus",
+        "multilogger"
+    ],
+    "cSpell.ignoreWords": [
+        "acarl",
+        "sirupsen"
+    ]
+}
