+++ generic_hook.go
@@ -7,7 +7,7 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// GenericHook represents a hook that logs at a given level. This struct must be extended to implement the Fire func
+// GenericHook represents a hook that logs at a given level. This struct must be extended to implement the Fire func.
 type GenericHook struct {
 	Formatter    logrus.Formatter
 	MinimumLevel logrus.Level
@@ -24,19 +24,23 @@ func (hook *GenericHook) formatEntry(entry *logrus.Entry) ([]byte, error) {
 	return formatted, nil
 }
 
-// Levels returns the levels that should be handled by the hook
+// Levels returns the levels that should be handled by the hook.
 func (hook *GenericHook) Levels() []logrus.Level {
-	levels := []logrus.Level{}
-	for _, level := range logrus.AllLevels {
-		levels = append(levels, level)
-		if level == hook.MinimumLevel {
-			return levels
+	switch level := hook.MinimumLevel; {
+	case level == DisabledLevel:
+		return nil
+	case level <= logrus.TraceLevel:
+		return logrus.AllLevels[:level+1]
+	default:
+		result := append(make([]logrus.Level, 0, level), logrus.AllLevels...)
+		for i := logrus.TraceLevel; i < level; i++ {
+			result = append(result, i+1)
 		}
+		return result
 	}
-	return []logrus.Level{}
 }
 
-// GetColor returns an ANSI color formatting function for every logrus logging level
+// GetColor returns an ANSI color formatting function for every logrus logging level.
 func GetColor(level logrus.Level) func(format string, args ...interface{}) string {
 	switch level {
 	case logrus.DebugLevel, logrus.TraceLevel:
