+++ level_test.go
@@ -1,6 +1,7 @@
 package multilogger
 
 import (
+	"fmt"
 	"strconv"
 	"strings"
 	"testing"
@@ -10,8 +11,10 @@ import (
 )
 
 func TestParseLevel(t *testing.T) {
+	assert.Equal(t, DisabledLevel, MustParseLogLevel(nil))
+	assert.Equal(t, DisabledLevel, MustParseLogLevel(""))
 	assert.Equal(t, DisabledLevel, MustParseLogLevel(DisabledLevel))
-	assert.Equal(t, DisabledLevel, MustParseLogLevel(DisabledLevelName))
+	assert.Equal(t, DisabledLevel, MustParseLogLevel(disabledLevelName))
 	assert.Equal(t, DisabledLevel, MustParseLogLevel(int(DisabledLevel)))
 	assert.Equal(t, DisabledLevel, MustParseLogLevel(strconv.Itoa(int(DisabledLevel))))
 	for _, level := range logrus.AllLevels {
@@ -24,6 +27,8 @@ func TestParseLevel(t *testing.T) {
 		assert.Equal(t, level, MustParseLogLevel(int(level)))
 		assert.Equal(t, level, MustParseLogLevel(strconv.Itoa(int(level))))
 	}
+	assert.Panics(t, func() { MustParseLogLevel("Invalid") })
+	assert.Panics(t, func() { MustParseLogLevel(1.23) })
 }
 
 func TestGetAcceptedLevels(t *testing.T) {
@@ -37,4 +42,19 @@ func TestParseInvalidLogLevel(t *testing.T) {
 	level, err = ParseLogLevel(1.234)
 	assert.Equal(t, level, DisabledLevel)
 	assert.EqualError(t, err, `Unable to parse the given logging level 1.234. It has to be a string or an integer`)
+}
+
+func xExample() {
+	logger, _ := Create(7, "debug", "", "test")
+	logger.SetReportCaller(true)
+	logger.Warn("Hello1")
+	logger.ConfigureFileLogger("debug", "tata")
+	logger.Error("This is an error message")
+	logger.Warn("This is a warning message")
+	logger.Info("This is an information message")
+	logger.Debug("This is a debug message")
+	logger.Trace("This is a trace message")
+	logger.WithField("hello", 1).Warn("This message has context")
+	fmt.Println(logger.consoleHook.Levels())
+	// Output:
 }
